Minor Refactoring to Split Camera and Controls
