+++ Hush/Assets/Scripts/Network.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e45249b597cc2c428e84618df447dd1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
