+++ Hush/Assets/Scripts/Player/Enums/PlayerAnimator.cs
